+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## WIP
+- add `lock` and `lock_stake` tx types
+
 ## 0.22.0 - 2021.12.01
 - add `add_limit_order`, `remove_limit_order` tx types